Update - fixes formatting, naming and commenting issues outlined in BESS feedback
