+++ src/Core/Block/Adminhtml/System/Config/Form/DeveloperLogs.php
@@ -23,7 +23,7 @@ use Magento\Framework\App\Filesystem\DirectoryList;
 
 /**
  * Class DeveloperLogs
- * @package Block\Adminhtml\System\Config\Form
+ * Displays links to available custom logs
  */
 class DeveloperLogs extends \Magento\Config\Block\System\Config\Form\Field
 {
