+++ src/Payment/view/frontend/web/js/view/notification.js
@@ -37,4 +37,3 @@ define(
         );
     }
 );
-
